Split Reconciler module into multiple files
